requirements.txt

apache-airflow-providers-apache-spark==4.0.0
requests
pandas
